correcciones

+++ pet/src/main/java/com/pethistory/pet/models/Rol.java
@@ -12,6 +12,6 @@ import lombok.Getter;
 public class Rol {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idRol;
+    private Long idRol;
     private String nomRol;
 }
